# backend/Services/PunchDataService/internal/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Collects raw punch logs from biometric devices, file uploads, manual entry, and read-only import from a public ZKTeco SQL Server. Stores normalised rows in PunchRecords with deduplication. Does not calculate attendance (see AttendanceService).",
        "title": "PunchDataService API",
        "contact": {
            "name": "EnterpriseERP Platform Team"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/punch-data/imports": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imports"
                ],
                "summary": "List import batch history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by company id (from query or JWT)",
                        "name": "companyId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by batch status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-github_com_enterprise-erp_punchdata_internal_response_PagedResult-github_com_enterprise-erp_punchdata_internal_models_PunchImportBatch"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/punch-data/imports/{id}/errors": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "imports"
                ],
                "summary": "List row errors for an import batch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Import batch id (uuid)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-github_com_enterprise-erp_punchdata_internal_response_PagedResult-github_com_enterprise-erp_punchdata_internal_models_PunchImportError"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_enterprise-erp_punchdata_internal_models.PunchImportBatch": {
            "type": "object",
            "properties": {
                "companyId": {
                    "type": "integer"
                },
                "contentType": {
                    "type": "string"
                },
                "duplicateRows": {
                    "type": "integer"
                },
                "errorMessage": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "insertedRows": {
                    "type": "integer"
                },
                "invalidRows": {
                    "type": "integer"
                },
                "processedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "totalRows": {
                    "type": "integer"
                },
                "uploadedAt": {
                    "type": "string"
                },
                "uploadedBy": {
                    "type": "string"
                },
                "validRows": {
                    "type": "integer"
                }
            }
        },
        "github_com_enterprise-erp_punchdata_internal_models.PunchImportError": {
            "type": "object",
            "properties": {
                "companyId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "errorMessage": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "importBatchId": {
                    "type": "string"
                },
                "rawRow": {
                    "type": "string"
                },
                "rowNumber": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_enterprise-erp_punchdata_internal_response.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiError"
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_enterprise-erp_punchdata_internal_response.ApiResponse-github_com_enterprise-erp_punchdata_internal_response_PagedResult-github_com_enterprise-erp_punchdata_internal_models_PunchImportBatch": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.PagedResult-github_com_enterprise-erp_punchdata_internal_models_PunchImportBatch"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiError"
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
        "github_com_enterprise-erp_punchdata_internal_response.ApiResponse-github_com_enterprise-erp_punchdata_internal_response_PagedResult-github_com_enterprise-erp_punchdata_internal_models_PunchImportError": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.PagedResult-github_com_enterprise-erp_punchdata_internal_models_PunchImportError"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiError"
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_enterprise-erp_punchdata_internal_response.PagedResult-github_com_enterprise-erp_punchdata_internal_models_PunchImportBatch": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_models.PunchImportBatch"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
        "github_com_enterprise-erp_punchdata_internal_response.PagedResult-github_com_enterprise-erp_punchdata_internal_models_PunchImportError": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_models.PunchImportError"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT issued by AuthService. Send as: `Bearer \u003ctoken\u003e`. Pass companyId in body or query when needed.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}